feat: simplify loading of required tiff files
