+++ packages/lambda-xyz/src/tiler.ts
@@ -1,6 +1,14 @@
 import { Tiler } from '@basemaps/tiler';
 
+/**
+ * Most tiles will be created by either a 256x256 tile or 512x512 tile
+ *
+ * This class is to cache the creation of the tilers, while also providing access
+ * so that they can be mocked during tests.
+ */
 export const Tilers = {
+    /** 256x256 Tiler */
     tile256: new Tiler(256),
+    /** 512x512 Tiler */
     tile512: new Tiler(512),
 };
